+++ programs/winewayland-host/unixlib.h
@@ -41,7 +41,7 @@ struct winewayland_host_probe
     char endpoint_path[WINEWAYLAND_HOST_NAME_MAX];
 };
 
-#define WINEWAYLAND_HOST_STARTUP_VERSION 9
+#define WINEWAYLAND_HOST_STARTUP_VERSION 10
 
 struct winewayland_host_startup
 {
@@ -75,6 +75,12 @@ struct winewayland_host_startup
     volatile uint32_t test_input_event_count;
     volatile uint32_t test_input_last_status;
     volatile uint32_t test_input_success_count;
+    volatile uint32_t empty_frames_presented;
+    volatile uint32_t empty_frame_width;
+    volatile uint32_t empty_frame_height;
+    volatile uint32_t frame_snapshot_test_status;
+    volatile uint32_t frame_snapshot_handles_before;
+    volatile uint32_t frame_snapshot_handles_after;
 };
 
 #define WINEWAYLAND_HOST_RENDERER_VERSION 16